Remove useless benchmark
